requirements.txt
requests
PyYAML
pytz
astral
tenacity
APScheduler
